Database implementation complete, Calculation tests are passing, Will have to verify the calculations more for accuracy. TODO: More tests TODO: Test organization
Amend: Reformatted code

+++ budget_manager_ui/src/expense_category/expense_category_object/mod.rs
@@ -7,7 +7,7 @@ glib::wrapper! {
 }
 
 impl ExpenseCategoryObject {
-    pub fn new(name: String, allocated: f32) -> Self {
+    pub fn new(name: String, allocated: f64) -> Self {
         Object::builder()
             .property("name", &name)
             .property("allocated", &allocated)
